+++ internal/cr_validation_test.go
@@ -33,6 +33,7 @@ import (
 	apiextensions "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	"sigs.k8s.io/yaml"
 
+	analyticsApi "github.com/arangodb/kube-arangodb/pkg/apis/analytics/v1alpha1"
 	appsv1 "github.com/arangodb/kube-arangodb/pkg/apis/apps/v1"
 	backupv1 "github.com/arangodb/kube-arangodb/pkg/apis/backup/v1"
 	deploymentv1 "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
@@ -223,6 +224,13 @@ func Test_GenerateCRValidationSchemas(t *testing.T) {
 			fmt.Sprintf("%s/pkg/apis/ml/v1alpha1", root): {
 				"v1alpha1": {
 					mlApiv1alpha1.ArangoMLBatchJob{}.Spec,
+				},
+			},
+		},
+		"analytics-graphanalyticsengine": {
+			fmt.Sprintf("%s/pkg/apis/analytics/v1alpha1", root): {
+				"v1alpha1": {
+					analyticsApi.GraphAnalyticsEngine{}.Spec,
 				},
 			},
 		},