+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - (Feature) (ML) Enable TLS
 - (Feature) (ML) Release V1Beta1 API
 - (Maintenance) Update Go to 1.22.3
+- (Feature) (Analytics) GAE Integration
 
 ## [1.2.40](https://github.com/arangodb/kube-arangodb/tree/1.2.40) (2024-04-10)
 - (Feature) Add Core fields to the Scheduler Container Spec